Revert "fix: don't use window in created (#59)"
This reverts commit 9b47c6ea630e35c8958a62f5b645fd3ef7216846.

Put the storage watcher back to the created() hook.  This fixes the
regression that breaks cross-window dark mode synchronization.

This bug is caused by the mounted() hook never being invoked, since the
$colorMode component is not actually mounted anywhere.  Also note that
there's nothing wrong in referencing "window" in the created() hook;
this is a client plugin, and other code in the same file references
"window" already.